Updated repository structure Tidied up Palm Beach README Added to gitignore
Updated repository structure
