+++ USA/FL/Palm_Beach/Court/README.md
@@ -1,12 +1,19 @@
-- install node
-- extract these files into a folder
-- navigate to that folder via terminal
-- type "npm install"
+# Florida: Palm Beach County
+
+This Scraper was built in JS before Python was decided as a project language.
+
+## Installation
+
+- Install node
+- `cd scraper`
+- `npm install`
 
 That should install all the node npm packages needed to run the script. You can modify the settings in r.js./ line 51
 
 Setting start to 0 should start you on the current day. Setting end to 1 should collect 2 days (day 0 & 1). I believe setting pages sets a max page count. There can be multiple pages per day.
 
-- type "npm start" 
+## Running
+- `cd scraper`
+- `npm start` 
 
-that will run the scraper in a custom chrome browser.
+This will run the scraper in a custom Chrome browser.
